Enabled TMC debug
Got some problem with uart mode.
The problem is that in S6 the PDN_UART pin is connected to pin 5 of stepper header but
in my stepper driver (BTT tmc2209 v1.1) the UART is connected to pin4 by default.
So instead of soldering a bridge on the driver I've added an external jumper between pin 4 and 5.
Now they works correctly!